Test to remove the links to the images.
The images of the code that I have learned and support currently have links embedded where the src is also the href, and for some reason when clicking on them they would open a new tab to the largest size of the images. This fix is to hopefully fix this.